refactor: modify signIn api
